pyproject.toml
[project]
name = "cuckoo-echo"
version = "0.1.0"
description = "Enterprise AI Customer Service SaaS Platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "granian>=1.6.0",
    "uvicorn[standard]>=0.30.0",
    "asyncpg>=0.30.0",
    "redis[hiredis]>=5.0.0",
    "pymilvus>=2.5.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    # "langgraph-store-postgres>=0.1.0",  # Not yet published; re-enable when available
    "litellm>=1.40.0,<1.80.0",  # Use older version without problematic tokenizers
    "langfuse>=2.40.0",
    "sentence-transformers>=3.0.0",
    "FlagEmbedding>=1.3.5,<1.3.8",  # Use version range that might have fixed the issue
    "sse-starlette>=2.0.0",
    "circuitbreaker>=2.0.0",
    "docling>=2.15.0",
    "docling-core>=2.3.0",
    "orjson>=3.10.0",
    "structlog>=24.4.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "minio>=7.2.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "PyJWT>=2.8.0",
    "bcrypt>=4.1.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.0",
    "psycopg[binary]>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-timeout>=2.3.0",
    "hypothesis>=6.100.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: requires running infrastructure (PostgreSQL, Milvus, Redis)",
    "e2e: end-to-end smoke tests requiring docker compose infrastructure",
]
filterwarnings = [
    # unittest.mock.AsyncMock creates coroutines that may be GC'd without awaiting
    # when MagicMock auto-creates AsyncMock children. This is a known CPython issue,
    # not a real bug in our code. See: https://github.com/python/cpython/issues/100226
    "ignore::RuntimeWarning:unittest.mock",
    # PBT tests use asyncio.new_event_loop() in sync wrappers for Hypothesis compatibility
    "ignore::DeprecationWarning:asyncio",
    # Suppress unclosed event loop warnings from PBT sync wrappers
    "ignore::pytest.PytestUnraisableExceptionWarning",
]

[tool.ruff]
target-version = "py311"
line-length = 120
